feat(schema): bigint 표준화 + Flyway V1 통합 + 이중 스키마 검증 CI
운영 스키마의 id 타입 혼재(int/bigint)와 FK 타입 미스매치 8건을 해소하고, 코드·DB·마이그레이션을 모두 bigint로 정렬한다.

- db/migration: V2~V11을 V1__baseline.sql로 통합. 모든 id PK=bigserial, FK=bigint (카운터 loop_count/version/block_index/prompt_tokens/completion_tokens는 integer 유지)

- SpringBoot 엔티티: HitlReview(존재하지 않던 target_id/review_prompt/review_comment 제거, comment/agenda_id/report_id 정렬), TokenUsageLog.cost→estimated_cost_usd(Double), ReportScore.total_score Float→Double

- FastAPI models.py: id/FK·recording_seconds=BigInteger, rating=SmallInteger

- CI(schema-validate.yml): 빈 postgres에 Flyway V1 적용 후 Hibernate validate + SQLAlchemy reflect로 양쪽 코드↔스키마 일치 검증

- ops/: 운영 일회성 정비 SQL(integer→bigint ALTER 42개 + Flyway 이력 리셋) + 런북

+++ backend/springboot/src/main/java/com/workmaite/domain/reports/entity/ReportScore.java
@@ -26,7 +26,7 @@ public class ReportScore {
     private String aiStatus = "pending";
 
     @Column(name = "total_score")
-    private Float totalScore;
+    private Double totalScore;
 
     @Column(name = "detail_scores", columnDefinition = "jsonb")
     private String detailScores;
@@ -38,7 +38,7 @@ public class ReportScore {
     @Column(name = "created_at", updatable = false)
     private LocalDateTime createdAt;
 
-    public void saveResult(String aiStatus, Float totalScore, String detailScores, String feedback) {
+    public void saveResult(String aiStatus, Double totalScore, String detailScores, String feedback) {
         this.aiStatus = aiStatus;
         this.totalScore = totalScore;
         this.detailScores = detailScores;
